+++ frontend/src/tsx/app.tsx
@@ -15,7 +15,7 @@ import '../css/layout.css?global';
 import '../css/global.css?global';
 
 import { UserState, BottomPanelState, AllChatState, EditorPanelState, MainScrollState } from './global_state';
-import { MainContent } from './main_content';
+import { BoardList } from './board_list';
 import { SignupPage } from './signup_page';
 import { InvitePage } from './invite_page';
 import { UserPage } from './profile/user_page';
@@ -34,6 +34,9 @@ function App(): JSX.Element {
 		let { setEmitter } = MainScrollState.useContainer();
 		return <div className="mainBody" ref={ref => setEmitter(ref)}>
 			<Switch>
+				<Route exact path="/app" render={() => (
+					<BoardList></BoardList>
+				)} />
 				<Route path="/app/register/:invite_code" render={props =>
 					<SignupPage {...props} />
 				} />
@@ -49,18 +52,11 @@ function App(): JSX.Element {
 				<Route path="/app/a/:article_id" render={props =>
 					<ArticlePage {...props} />
 				} />
+				<Route path="/app/b/:board_name" render={() =>
+					<BoardSwitch />
+				} />
 				<Route path="*" render={() =>
-					<div className="forumBody">
-						<Switch>
-							<Route exact path="/app" render={() => (
-								<MainContent></MainContent>
-							)} />
-							<Route path="/app/b/:board_name" render={() =>
-								<BoardSwitch />
-							} />
-							<Redirect to="/app" />
-						</Switch>
-					</div>
+					<Redirect to="/app" />
 				} />
 			</Switch>
 		</div>;
